+++ cdm-web/src/main/java/org/cdm/web/backend/CdmWebApplication.java
@@ -1,4 +1,4 @@
-package ru.nsu.cdm.cdmweb;
+package org.cdm.web.backend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
