+++ cdm-web/src/test/java/org/cdm/web/backend/CdmWebApplicationTests.java
@@ -1,4 +1,4 @@
-package ru.nsu.cdm.cdmweb;
+package org.cdm.web.backend;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
